Sort DK runes left to right according to their CD
Some addons (e.g https://wow.curseforge.com/projects/onerunebar) allow
treating runes like a single energy bar, sorting the most full runes to
the left and the emptiest runes to the right. The visual effect is quite
nice and helpful.

This is an initial implementation of the behaviour described above, for
now without configurable options.